# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'API Recette'
  description: 'Cette API participative permet de partager des recettes entre utilisateurs.'
  version: '1'
servers:
  -
    url: 'http://localhost:8000'
    description: 'Mon API'
paths:
...
  '/users/deactivate/{id}':
    get:
      tags:
        - User
      summary: 'Deactivate an User'
      description: 'Deactivate an User'
      operationId: 'App\Controller\UserController::deactivate'
      parameters:
        -
          name: id
          in: path
          description: 'id de l''utilisateur'
          required: true
      responses:
        '200':
          description: 'Deactivate a User'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'L''utilisateur a bien été désactivé.' }
                type: object
        '404':
          $ref: '#/components/responses/notFound'
        '403':
          $ref: '#/components/responses/forbidden'
  '/users/activate/{id}':
    get:
      tags:
        - User
      summary: 'Activate an User'
      description: 'Activate an User'
      operationId: 'App\Controller\UserController::activate'
      parameters:
        -
          name: id
          in: path
          description: 'id de l''utilisateur'
          required: true
      responses:
        '200':
          description: 'Activate a User'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'L''utilisateur a bien été activé.' }
                type: object
        '404':
          $ref: '#/components/responses/notFound'
        '403':
          $ref: '#/components/responses/forbidden'
components:
...
  responses:
    notFound:
      description: 'This resource not found'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Cette ressource n''existe pas'
            type: object
...
    forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Vous n''êtes pas autorisé à effectuer cette action.'
            type: object
...
  securitySchemes:
    basic:
      type: http
      scheme: basic